% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add data to a probability distribution object}
\usage{
add_data(d, data)
}
\arguments{
\item{d}{A probability distribution object such as those created by a call
to \code{\link[distributions3:Normal]{Normal()}} or
\code{\link[distributions3:Binomial]{Binomial()}} etc.}

\item{data}{A numeric vector, matrix, data frame or list providing a numeric
vector of data for each of the \code{length(d)} distinct variable in \code{d}.
\code{data} may be a matrix with \code{length(d)} columns, a data frame with
\code{length(d)} variables or a list of length \code{length(d)}. A matrix or data
frame may contain \code{NA} values. Only if \code{length(d) = 1} may \code{data} be a
numeric vector.}
}
\value{
An object with the same class as the input \code{d}, that is, inheriting
from class \code{"distribution"}.
}
\description{
Describe
}
\details{
Additional details... Explain that all the data must be in the
support of distribution \code{d}.
}
\examples{
library(distributions3)

# Note: the value of p is irrelevant
x <- Binomial(size = 10)
x <- add_data(x, 1:3)
y <- Beta(alpha = 1, beta = 1)
z <- x * y
z

x <- Binomial(size = c(5, 10))
data <- list(1:2, 1:3)
x <- add_data(x, data)
y <- Beta(alpha = 1, beta = 1)
z <- x * y
z

}
\seealso{
\code{\link{posterior}} for performing Bayesian inference using
\code{"distribution"} objects.
}
