% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDIbinomial.R
\name{MDIbinomial}
\alias{MDIbinomial}
\title{Create an MDI prior distribution for a binomial probability}
\usage{
MDIbinomial(alpha = 1, beta = 1)
}
\arguments{
\item{alpha, beta}{Non-negative shape parameters \eqn{\alpha} and \eqn{\beta}.}
}
\value{
A \code{MDIbinomial} object, inheriting from class \code{"distribution"}.
}
\description{
Creates a distribution with p.d.f.
\eqn{f_X(x) \propto x^{\alpha x} (1-x)^{\beta(1-x)}} for \eqn{0 < x < 1}.
The case \eqn{\alpha = \beta = 1} arises as the Maximal Data Information
(MDI) prior for the probability parameter \eqn{p} of a Binomial distribution.
}
\details{
See Tuyl et al. (2008) for further information and a comparison
of the MDI prior with other prior distributions often used for Binomial
data.
}
\examples{
P <- MDIbinomial()
P
}
\references{
Tuyl, F., Gerlach, R., and Mengersen, K. (2008). A Comparison of
Bayes–Laplace, Jeffreys, and Other Priors: The Case of Zero Events.
\emph{The American Statistician}, \strong{62}(1), 40–44.
\doi{10.1198/000313008X267839}
}
\seealso{
\code{\link{posterior}} for an example where this distribution is used as a
(non-conjugate) prior for the probability parameter \eqn{p} of a Binomial
distribution.
}
