% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalGamma.R
\name{NormalGamma}
\alias{NormalGamma}
\title{Create a Normal-Gamma distribution}
\usage{
NormalGamma(mu = 0, lambda = 1, shape = 1, rate = 1)
}
\arguments{
\item{mu, lambda}{Parameters \eqn{\mu} (location) and \eqn{\lambda}
(the precision is \eqn{\lambda Y}) for a Normal distribution.
See \code{\link[distributions3:Normal]{Normal()}}. \code{lambda} must be positive.}

\item{shape, rate}{Shape and rate parameters of a Gamma distribution. See
\code{\link[distributions3:Gamma]{Gamma()}}. \code{shape} and \code{rate} must be positive.}
}
\value{
A \code{NormalGamma} object, inheriting from class \code{"distribution"}.
}
\description{
Creates a bivariate distribution for \eqn{(X, Y)} where \eqn{X \mid Y} has a
Normal distribution with mean \eqn{\mu} and precision \eqn{\lambda Y}
(variance \eqn{1/\lambda Y}), and \eqn{Y} has a Gamma distribution.
}
\details{
See the Wikipedia page
\href{https://en.wikipedia.org/wiki/Normal-gamma_distribution}{Normal-gamma distribution}.
See \code{\link[distributions3:Normal]{Normal()}} and
\code{\link[distributions3:Gamma]{Gamma()}} for information about the Normal and
Gamma distributions. The joint p.d.f. of \eqn{(X, Y)} is the product of
the conditional (normal) p.d.f. of \eqn{X \mid Y} and the (gamma) p.d.f.
of \eqn{Y}.
}
\examples{
Z <- NormalGamma()
Z
}
\seealso{
\code{\link{posterior}} for an example where a Normal-Gamma distribution is
used as a conjugate prior for the location and precision parameters of a
Normal distribution.
}
