% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDIbinomial.R
\name{pdf.MDIbinomial}
\alias{pdf.MDIbinomial}
\alias{log_pdf.MDIbinomial}
\title{Evaluate the probability density function of an MDI prior distribution for a
binomial probability}
\usage{
\method{pdf}{MDIbinomial}(d, x, drop = TRUE, elementwise = NULL, ...)

\method{log_pdf}{MDIbinomial}(d, x, drop = TRUE, elementwise = NULL, ...)
}
\arguments{
\item{d}{An \code{MDIbinomial} object created by a call to \code{\link[=MDIbinomial]{MDIbinomial()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{d} be evaluated
at all elements of \code{x} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{d} and \code{x} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{...}{Not used.}
}
\value{
In case of a single distribution object, either a numeric
vector of length \code{probs} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(x)} columns (if \code{drop = FALSE}). In case of a vectorized
distribution object, a matrix with \code{length(x)} columns containing all
possible combinations.
}
\description{
Evaluate the probability density function of an MDI prior distribution for a
binomial probability
}
