% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/Binomial.R, R/Normal.R
\name{*.distribution}
\alias{*.distribution}
\alias{posterior}
\alias{posterior.Binomial}
\alias{posterior.Normal}
\title{Create a posterior distribution}
\usage{
\method{*}{distribution}(x, y)

posterior(x, y, ...)

\method{posterior}{Binomial}(x, y, ...)

\method{posterior}{Normal}(x, y, ...)
}
\arguments{
\item{x}{To specify a \strong{likelihood}. A probability distribution object
inheriting from class \code{"distribution"} such as those created by a call to
\code{\link[distributions3:Normal]{Normal()}} or
\code{\link[distributions3:Binomial]{Binomial()}} etc. \code{x} must have an extra
attribute \code{"data"}, added using \code{\link[=add_data]{add_data()}}.}

\item{y}{To specify a \strong{prior} distribution. A probability distribution
object inheriting from class \code{"distribution"}.}

\item{...}{Further arguments to be passed to \code{\link[rust:ru]{ru()}}. See
\strong{Details}.}
}
\value{
In conjugate cases, a probability distribution object from the same
family as the prior distribution. The first component of the class of this
object is the name of the prior (and posterior) distribution.

In non-conjugate cases, an object, of class \code{c("ru", "posterior")},
returned from \code{\link[rust:ru]{ru()}}.
}
\description{
Generic functions for performing univariate Bayesian inference using
\code{"distribution"} objects. A generic multiplication operator \verb{*.distribution}
creates a posterior distribution as a product of a likelihood and
prior distribution. Alternatively, \code{posterior(x, y)} or can be used.
}
\details{
If \code{x} has an attribute \code{"data"}, added using \code{\link[=add_data]{add_data()}}, then
\code{x} is used to define a likelihood based on treating \code{attr(x, "data")} as
a random sample from the probability distribution underlying \code{x}. All
observations in the data must be in the support of the probability
distribution \code{d}. See \code{\link[distributions3:support]{support()}}. The
posterior distribution for the unknown parameters of \code{x} is inferred using
the prior distribution supplied by \code{y}. \code{posterior(x, y)} will give the
same output.

A message will note likelihood and prior combination used to create the
posterior distribution. These messages can be suppressed using
\code{\link[=message]{suppressMessages()}} if desired.

\strong{Conjugate priors}. For some likelihoods, it is possible to specify a
family of prior distribution such that the posterior distribution is in
the same family. For example, for a likelihood based on a random sample
from a \code{\link[distributions3:Binomial]{Binomial()}} distribution, specifying
a \code{\link[distributions3:Beta]{Beta()}} prior distribution for the
probability parameter \eqn{p} results in a Beta posterior distribution.
In cases like this, \code{posterior()} returns a distribution object with the
same distribution type as the distribution object \code{y} used to specify
the prior distribution. See the table of conjugate distributions in the
wikipedia article
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{Conjugate prior}
for details and examples of other conjugate cases.

\strong{Non-conjugate priors}. If \code{y} is not a conjugate prior distribution for
the likelihood \code{x}, then \code{\link[rust:ru]{ru()}} is used to simulate from the
posterior distribution. In this event \code{...} may be used in a call to
\code{posterior(x, y, ...)} to pass arguments to \code{\link[rust:ru]{ru()}}. The
argument \code{n} may be useful because it specifies the size of the posterior
sample. The default set in the \code{posterior} methods supplied in
\code{posteriors3} is \code{n = 1000}. Other arguments to \code{\link[rust:ru]{ru()}} have
default settings designed to work well in specific examples. For example,
in \code{\link[=posterior]{posterior.Binomial()}} sampling from the posterior for
the probability parameter \eqn{p} is conducted on a logit scale, that is,
we sample from the posterior distribution for \eqn{\log(p / (1-p))}.
See \strong{Examples} for an example based on the (non-conjugate) Maximal Data
Information (MDI) prior \code{\link[=MDIbinomial]{MDIbinomial()}}.
}
\examples{
library(distributions3)

### Binomial likelihood, conjugate Beta prior for p

## One Binomial distribution, with size 10

# Note: the value of p is only use to simulate example data
N <- Binomial(size = 10, p = 0.2)

# Simulate a sample of size 5 from a Binomial(20, 0.2) distribution
set.seed(3)
data <- random(N, 5)

# Add the data, a numeric vector
likelihood <- add_data(N, data)

# Set a conjugate (uniform) prior distribution
prior <- Beta(alpha = 1, beta = 1)
# Construct the posterior distribution
posterior <- likelihood * prior
posterior
plot(posterior)

# Compare two different conjugate prior distributions
prior <- Beta(alpha = c(1, 10), beta = c(1, 10))
# Determine the posterior distribution
posterior <- likelihood * prior
posterior
plot(posterior)

# Non-conjugate MDI prior, simulation from the posterior
prior <- MDIbinomial()
posterior <- likelihood * prior
plot(posterior)

## Two Binomial distributions, with sizes 5 and 10

# Note: the value of p is only use to simulate example data
M <- Binomial(size = c(5, 10), p = 0.8)

# Simulate samples of size 8 from
#   a Binomial(5, 0.8) distribution, and
#   a Binomial(10, 0.8) distribution

set.seed(3)
data <- random(M, 8)
# data is a matrix with 2 rows: one per Binomial distribution

# Below are two more equivalent ways to supply these data

# Extract the individual samples
data1 <- data[1, ]
data2 <- data[2, ]
# data is a list of length 2: one numeric vector per Binomial distribution
data <- list(data1 = data1, data2 = data2)
# data is a data frame with 2 variables: one per Binomial distribution
data <- data.frame(data1, data2)

likelihood <- add_data(M, data)
prior <- Beta(alpha = 1.5, beta = 2.1)
posterior <- likelihood * prior
posterior
plot(posterior, legend_args = list(x = "topleft"))

### Normal likelihood, conjugate (normal, gamma) prior for (mu, 1/sigma^2)

# Note: the values of mu and sigma are only used to simulate example data
X <- Normal(mu = 10, sigma = 2)

# Simulate a sample of size 15 from a Normal(10, 2^2) distribution
set.seed(3)
data <- random(X, 15)
# Add the data, a numeric vector
likelihood <- add_data(X, data)

# Set a Normal-Gamma prior for (mean mu, precision tau)
prior <- NormalGamma(names = c("mu", "tau"))
posterior <- likelihood * prior
posterior

# Plot the marginal prior and posterior distributions for mu and tau
plot(posterior, margin = "mu")
plot(posterior, margin = "tau")

# Find the marginal posterior distributions for mean and precision
x <- marginals(posterior)
# plot the marginal posterior distributions individually
plot(x[[1]])
plot(x[[2]])

# Contour plots of posterior and prior distributions
plot(posterior)
}
\seealso{
\code{\link[=add_data]{add_data()}} for adding data to a \code{"distribution"} object.

\code{\link[=plot.posterior]{plot.posterior()}} for plotting posterior and prior distributions.
}
