% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalGamma.R
\name{random.NormalGamma}
\alias{random.NormalGamma}
\title{Draw a random sample from a Normal-Gamma distribution}
\usage{
\method{random}{NormalGamma}(x, n = 1L, drop = TRUE, ...)
}
\arguments{
\item{x}{A \code{NormalGamma} object created by a call to \code{\link[=NormalGamma]{NormalGamma()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n}
columns (if \code{drop = FALSE}). If \code{x} contains more than one distribution
then a matrix with \code{n} columns is returned, with each row containing a
random sample from one of the distributions.
}
\description{
Draw a random sample from a Normal-Gamma distribution
}
\examples{
X <- NormalGamma(mu = 10, lambda = 1, shape = 5, rate = 1)
random(X, 10)

}
