% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.R
\name{marginals}
\alias{marginals}
\alias{marginals.NormalGamma}
\title{Extract marginal distributions}
\usage{
marginals(d, ...)

\method{marginals}{NormalGamma}(d, ...)
}
\arguments{
\item{d}{A probability distribution object inheriting from class
\code{"distribution"} such as those created by a call to \code{\link[=NormalGamma]{NormalGamma()}}.}

\item{...}{Not used.}
}
\value{
A list with \code{length(d)} components. Each component is a univariate
distribution object inheriting from class \code{"distribution"}.
}
\description{
Extract marginal distributions from a multivariate distribution inheriting
from class \code{"distribution"}.
}
\details{
Additional details...
}
