% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalGamma.R
\name{pdf.NormalGamma}
\alias{pdf.NormalGamma}
\alias{log_pdf.NormalGamma}
\title{Evaluate the joint probability density function of a Normal-Gamma distribution}
\usage{
\method{pdf}{NormalGamma}(d, x, drop = TRUE, ...)

\method{log_pdf}{NormalGamma}(d, x, drop = TRUE, ...)
}
\arguments{
\item{d}{A \code{NormalGamma} object created by a call to \code{\link[=NormalGamma]{NormalGamma()}}.}

\item{x}{A numeric matrix with 2 columns. Column 1 provides the values for
the \eqn{X \mid Y} component and column 2 the corresponding values of the
\eqn{Y} components of the joint distribution of \eqn{(X, Y)}, where
\eqn{X \mid Y} has a Normal distribution with mean \eqn{\mu} and
precision \eqn{\lambda Y} and has a Gamma distribution.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Not used.}
}
\value{
If \code{d} is a single distribution object, either a numeric
vector of length \code{max(length(x), length(y))} (if \code{drop = TRUE}, default)
or a \code{matrix} with \code{length(x)} columns (if \code{drop = FALSE}). In case of a
vectorized distribution object, a matrix with \code{length(x)} columns, one for
each distribution.
}
\description{
Evaluate the joint probability density function of a Normal-Gamma distribution
}
\examples{
library(distributions3)
X <- NormalGamma()
pdf(X, matrix(c(0, 0), ncol = 2))
}
