% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/Binomial.R
\name{*.distribution}
\alias{*.distribution}
\alias{posterior}
\alias{posterior.Binomial}
\title{Posterior distributions}
\usage{
\method{*}{distribution}(x, y)

posterior(x, y)

\method{posterior}{Binomial}(x, y)
}
\arguments{
\item{x}{To specify the \strong{likelihood}. A probability distribution object
inheriting from class \code{"distribution"} such as those created by a call to
\code{\link[distributions3:Normal]{Normal()}} or
\code{\link[distributions3:Binomial]{Binomial()}} etc. \code{x} must have a data
attribute, added using \code{\link[=add_data]{add_data()}}.}

\item{y}{To specify the \strong{prior} distribution. A probability distribution
object inheriting from class \code{"distribution"}.}
}
\value{
A probability distribution object with the same class as the prior
distribution object \code{y}.
}
\description{
Generic functions for performing Bayesian inference using \code{"distribution"}
objects. After using \code{\link[=add_data]{add_data()}} to add a random sample of data to a
distribution object \code{x} the generic multiplication \verb{*.distribution}
multiplies the likelihood by the prior distribution specified in \code{y} to
create a posterior distribution.
}
\details{
Repeat the explanation in \code{add_data} concerning the data being in
the support of \code{x}. Explain conjugacy and give an example (Binomial-Beta).
If prior is conjugate then ... distribution object: same distribution type
as the prior distribution in \code{y}.
If not then ... Sample from posterior.
}
\examples{
library(distributions3)

# Note: the value of p is irrelevant
x <- Binomial(size = 10)
x <- add_data(x, 1:3)
# Using a conjugate prior distribution
y <- Beta(alpha = 1, beta = 1)
z <- x * y
z

# Two binomial distributions
x <- Binomial(size = c(5, 10))
# Two vectors of data
data <- list(1:2, 1:3)
x <- add_data(x, data)
y <- Beta(alpha = 2, beta = 2)
z <- x * y
z
}
\seealso{
\code{\link[=add_data]{add_data()}} for adding data to a \code{"distribution"} object.
}
