% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/Binomial.R
\name{*.distribution}
\alias{*.distribution}
\alias{posterior}
\alias{posterior.Binomial}
\title{Posterior distributions}
\usage{
\method{*}{distribution}(x, y)

posterior(x, y)

\method{posterior}{Binomial}(x, y)
}
\arguments{
\item{x}{To specify the \strong{likelihood}. A probability distribution object
inheriting from class \code{"distribution"} such as those created by a call to
\code{\link[distributions3:Normal]{Normal()}} or
\code{\link[distributions3:Binomial]{Binomial()}} etc. \code{x} must have an extra
attribute \code{"data"}, added using \code{\link[=add_data]{add_data()}}.}

\item{y}{To specify the \strong{prior} distribution. A probability distribution
object inheriting from class \code{"distribution"}.}
}
\value{
A probability distribution object with the same class as the prior
distribution object \code{y}.
}
\description{
Generic functions for performing Bayesian inference using \code{"distribution"}
objects. After using \code{\link[=add_data]{add_data()}} to add a random sample of data to a
distribution object \code{x} the generic multiplication \verb{*.distribution}
multiplies the implied likelihood in \code{x} by the prior distribution specified
in \code{y} to create a posterior distribution. \code{posterior(x, y)} will give the
same output.
}
\details{
Repeat the explanation in \code{add_data} concerning the data being in
the support of \code{x}. Explain conjugacy and give an example (Binomial-Beta).
If prior is conjugate then ... distribution object: same distribution type
as the prior distribution in \code{y}.
If not then ... Sample from posterior.
}
\examples{
library(distributions3)

## One Binomial distribution, with size 10
## Note: the argument p is irrelevant

distribution <- Binomial(size = 10)

# Add the data, a numeric vector c(1, 2, 3)
likelihood <- add_data(distribution, 1:3)

# Set a conjugate prior distribution
prior <- Beta(alpha = 1, beta = 1)
# Determine the posterior distribution
posterior <- likelihood * prior
posterior

## Two Binomial distributions, with sizes 5 and 10

distributions <- Binomial(size = c(5, 10))

# Below are three equivalent ways to supply the data
#   1,2 for the first Binomial distribution, and
#   1, 2, 3 for the second Binomial distribution

# data is a list of length 2: one numeric vector per Binomial distribution
data <- list(1:2, 1:3)
# data is a data frame with 2 variables: one per Binomial distribution
data <- data.frame(data1 = c(1:2, NA), data2 = 1:3)
# data is a matrix with 2 columns: one per Binomial distribution
data <- matrix(c(1:2, NA, 1:3), nrow = 3, ncol = 2)

likelihood <- add_data(distributions, data)
prior <- Beta(alpha = 1.5, beta = 2.1)
posterior <- likelihood * prior
posterior

}
\seealso{
\code{\link[=add_data]{add_data()}} for adding data to a \code{"distribution"} object.
}
