% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/Binomial.R, R/Normal.R
\name{*.distribution}
\alias{*.distribution}
\alias{posterior}
\alias{posterior.Binomial}
\alias{posterior.Normal}
\title{Create a posterior distribution}
\usage{
\method{*}{distribution}(x, y)

posterior(x, y)

\method{posterior}{Binomial}(x, y)

\method{posterior}{Normal}(x, y)
}
\arguments{
\item{x}{To specify a \strong{likelihood}. A probability distribution object
inheriting from class \code{"distribution"} such as those created by a call to
\code{\link[distributions3:Normal]{Normal()}} or
\code{\link[distributions3:Binomial]{Binomial()}} etc. \code{x} must have an extra
attribute \code{"data"}, added using \code{\link[=add_data]{add_data()}}.}

\item{y}{To specify a \strong{prior} distribution. A probability distribution
object inheriting from class \code{"distribution"}.}
}
\value{
A probability distribution object from the same family as the prior
distribution used.
}
\description{
Generic functions for performing univariate Bayesian inference using
\code{"distribution"} objects. A generic multiplication operator \verb{*.distribution}
creates a posterior distribution as a product of a likelihood and
prior distribution. Alternatively, \code{posterior(x, y)} or can be used.
}
\details{
If \code{x} has an attribute \code{"data"}, added using \code{\link[=add_data]{add_data()}}, then
\code{x} is used to define a likelihood based on treating \code{attr(x, "data")} as
a random sample from the probability distribution underlying \code{x}. All
observations in the data must be in the support of the probability
distribution \code{d}. See \code{\link[distributions3:support]{support()}}. The
posterior distribution for the unknown parameters of \code{x} is inferred using
the prior distribution supplied by \code{y}. \code{posterior(x, y)} will give the
same output.

A message will note likelihood and prior combination used to create the
posterior distribution. Such messages can be suppressed using
\code{\link[=message]{suppressMessages()}} if desired.

\strong{Explain conjugacy and give an example (Binomial-Beta?).
If prior is conjugate then ... distribution object: same distribution type
as the prior distribution in \code{y}.
If not then ... sample from posterior.}
}
\examples{
library(distributions3)

### Binomial likelihood, conjugate Beta prior for p

## One Binomial distribution, with size 10

# Note: the value of p is only use to simulate example data
N <- Binomial(size = 10, p = 0.2)

# Simulate a sample of size 5 from a Binomial(20, 0.2) distribution
set.seed(3)
data <- random(N, 5)

# Add the data, a numeric vector
likelihood <- add_data(N, data)

# Set a conjugate (uniform) prior distribution
prior <- Beta(alpha = 1, beta = 1)
# Construct the posterior distribution
posterior <- likelihood * prior
posterior
plot(posterior)

# Compare two different conjugate prior distributions
prior <- Beta(alpha = c(1, 10), beta = c(1, 10))
# Determine the posterior distribution
posterior <- likelihood * prior
posterior
plot(posterior)

## Two Binomial distributions, with sizes 5 and 10

# Note: the value of p is only use to simulate example data
M <- Binomial(size = c(5, 10), p = 0.8)

# Simulate samples of size 8 from
#   a Binomial(5, 0.8) distribution, and
#   a Binomial(10, 0.8) distribution

set.seed(3)
data <- random(M, 8)
# data is a matrix with 2 rows: one per Binomial distribution

# Below are two more equivalent ways to supply these data

# Extract the individual samples
data1 <- data[1, ]
data2 <- data[2, ]
# data is a list of length 2: one numeric vector per Binomial distribution
data <- list(data1 = data1, data2 = data2)
# data is a data frame with 2 variables: one per Binomial distribution
data <- data.frame(data1, data2)

likelihood <- add_data(M, data)
prior <- Beta(alpha = 1.5, beta = 2.1)
posterior <- likelihood * prior
posterior
plot(posterior, legend_args = list(x = "topleft"))

### Normal likelihood, conjugate (normal, gamma) prior for (mu, 1/sigma^2)

# Note: the values of mu and sigma are only used to simulate example data
X <- Normal(mu = 10, sigma = 2)

# Simulate a sample of size 15 from a Normal(10, 2^2) distribution
set.seed(3)
data <- random(X, 15)

# Add the data, a numeric vector
likelihood <- add_data(X, data)

# Set a Normal-Gamma prior for (mean mu, precision tau)

prior <- NormalGamma(names = c("mu", "tau"))
posterior <- likelihood * prior
posterior

# Plot the marginal prior and posterior distributions for mu and tau
plot(posterior, margin = "mu")
plot(posterior, margin = "tau")

# Find the marginal posterior distributions for mean and precision
x <- marginals(posterior)
# plot the marginal posterior distributions individually
plot(x[[1]])
plot(x[[2]])

# plot(posterior)
}
\seealso{
\code{\link[=add_data]{add_data()}} for adding data to a \code{"distribution"} object.

\code{\link[=plot.posterior]{plot.posterior()}} for plotting posterior and prior distributions.
}
