% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors3-internal.R
\name{posteriors3-internal}
\alias{posteriors3-internal}
\alias{check_data}
\alias{check_support}
\title{Internal posteriors3 functions}
\usage{
check_data(x)

check_support(x)
}
\description{
Internal posteriors3 functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
