% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors3-package.R
\docType{package}
\name{posteriors3-package}
\alias{posteriors3}
\alias{posteriors3-package}
\title{posteriors3: Univariate Bayesian Analyses using S3 Probability Distribution Objects}
\description{
Describe
}
\details{
Details
}
\references{
Hayes A, Moller-Trane R, Jordan D, Northrop P, Lang M, Zeileis A
(2024). distributions3: Probability Distributions as S3 Objects. R package
version 0.2.2, \doi{10.32614/CRAN.package.distributions3}
}
\seealso{
\code{\link[=add_data]{add_data()}} for adding data to a \code{"distribution"} object.

\code{\link{Bayesian}} for calculating products of functions for Bayesian
Inference using \code{"distribution"} objects.
}
\author{
\strong{Maintainer}: Paul J. Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

}
